G: Modification du type de nombre de int a float

+++ src/string_calculator.py
@@ -17,7 +17,7 @@ class StringCalculator :
         product = 1
         for part in parts :
             try :
-                number = int(part)
+                number = float(part)
             except ValueError:
                  raise ValueError("Invalid input: '{}' is not a valid number.".format(part))
             if number <= 1000 :
